表单改用弹窗

+++ src/components/Index.vue
@@ -1,15 +1,9 @@
 <template>
   <div class="index">
-    <el-menu theme="dark" default-active="1" mode="horizontal">
-      <el-menu-item index="1">项目管理</el-menu-item>
-      <el-menu-item index="1">建筑管理</el-menu-item>
-      <el-menu-item index="2">订单管理</el-menu-item>
-    </el-menu>
-
     <h1 class="title">所有项目</h1>
     <p class="btn-add">
       <el-button type="success" icon="plus"
-        @click="$router.push('/editor/projects/new')">添加项目</el-button>
+        @click="openDialog('create', '')">添加项目</el-button>
     </p>
 
     <el-table :data="tableData" border style="width: 100%">
@@ -22,32 +16,67 @@
       <el-table-column align="center" label="创建时间" prop="create_date" :formatter="fmtDate"/>
       <el-table-column align="center" label="操作">
         <template slot-scope="scope">
-          <el-button size="small" @click="$router.push(`/editor/projects/${scope.row.id}`)">编辑</el-button>
+          <el-button size="small" @click="openDialog('edit', scope.row.id)">编辑</el-button>
           <el-button size="small" type="danger" @click="handleDelete(scope.row.id)">删除</el-button>
         </template>
       </el-table-column>
     </el-table>
+
+    <el-dialog title="项目表单" :visible.sync="dialogFormVisible">
+      <projects-form :type="dialogFormType" :id="projectId" @close="closeDialog"/>
+    </el-dialog>
   </div>
 </template>
 
 <script>
   import axios from 'axios'
+  import projectsForm from '@/components/form/projectsForm'
 
   export default {
     name: 'Index',
     data () {
       return {
+        dialogFormVisible: false,
+        dialogFormType: 'create',
+        projectId: null,
         tableData: []
       }
     },
+    components: {
+      projectsForm
+    },
     methods: {
+      openDialog (type, projectId) {
+        this.dialogFormType = type
+        this.projectId = projectId
+        this.dialogFormVisible = true
+      },
+      closeDialog (flag) {
+        if (flag) {
+          this.fetchTableData()
+          this.$message({ type: 'success', message: '编辑成功' })
+        } else {
+          this.$message({ type: 'info', message: '已取消编辑' })
+        }
+        this.projectId = ''
+        this.dialogFormVisible = false
+      },
       fmtDate (row, column, dateStr) {
         return dateStr ? dateStr.replace(/(\d{4}-\d{2}-\d{2}).*?(\d{2}:\d{2}).*/, '$1 $2') : ''
       },
       handleDelete (id) {
-        axios.delete(`/api/projects/${id}`).then(res => {
-          this.fetchTableData()
-        }).catch(err => console.log(err))
+        this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          axios.delete(`/api/projects/${id}`).then(res => {
+            this.fetchTableData()
+            this.$message({ type: 'success', message: '删除成功' })
+          }).catch(err => console.log(err))
+        }).catch(() => {
+          this.$message({ type: 'info', message: '已取消删除' })
+        })
       },
       fetchTableData () {
         axios.get('/api/projects').then(res => {
